requirements.txt
python-dotenv
dspy
openai